refactor: move tests related to corrections
